src/slack-suppressor/requirements.txt
Flask
slackclient
